refactor: Replace AgentCard with SettingsCard in LibraryTab
