+++ apps/desktop/src/components/settings/agents/index.ts
@@ -1,4 +1,3 @@
-export { AgentCard } from "./AgentCard";
 export { AgentsSection } from "./AgentsSection";
 export { EmptyLibraryState } from "./EmptyLibraryState";
 export { EmptyTemplatesState } from "./EmptyTemplatesState";
